Add fix_backslashes()
To be clear, I'm not a fan of this -- I think that \\ in a man page is
much easier to understand than \e.  However, mandoc_char(7) is very
clear that \\ is "not the right way to output a backslashes".

Personally, I'm not convinced that their reasons are worth it.  But this
is an easy fix to make if it's desired by FreeBSD maintainers and
committers, so I figured I'd implement it and toss the fixes their way
for discussion.